requirements.txt
streamlit
pinecone
sentence-transformers
google-genai
langgraph
python-dotenv
pandas
numpy
pillow
